fix: 彻底弄清楚并解决了使用预激活时导致的entropy瞬间归零和alpha、alpha_loss值爆炸问题   - 使用预激活，每个Dense前必须用LN。而Actor和Critic中的Dense遗漏了

+++ BS_RL/SAC/demo_TradingEnv.py
@@ -12,7 +12,7 @@ if __name__ == "__main__":
     env_id = "TradingEnv"
     env_num = 96
     eval_env_num = 48
-    trading_env_config = TradingEnvConfig(data_path="/root/project/processed_data/",window_size_5m=4*60/5)
+    trading_env_config = TradingEnvConfig(data_path="/root/project/processed_data/",window_size_5m=int(4*60/5))
     is_test = total_timesteps!=int(200e6)
     learning_starts = int(batch_size) if is_test else int(2e4)
     ckpt_save_frequency = None if is_test else 0.01
